some code tweaks * simpler code * now tooltip with unread mail count is working * proper email id is passed

+++ IMAP.Popup/Models/Email.cs
@@ -16,6 +16,6 @@ namespace IMAP.Popup.Models
 
         public DateTime WhenSent { get; set; }
 
-        public long MessageUid { get; set; }
+        public uint MessageUid { get; set; }
     }
 }
